MI - total jobs by status for each accountId
