+++ test/unit/reports.spec.js
@@ -7,15 +7,18 @@ const reports = require('../../db/reports.js');
 describe('Reports', function () {
   describe('Metrics', function () {
     before(function () {
-      return knex.seed.run({ directory: './db/seeds/two-jobs-updated-over-a-week-ago' });
+      return knex.seed.run({ directory: './db/seeds/reports-total-saved-jobs' });
     });
 
-    it('should return total saved jobs per account', function () {
+    it('should return count of job statuses per account', () =>
       reports.getTotalSavedJobs()
-        .then((result) => {
-          expect(result).to.contain('interventionRef,totalSaved\nFILTER,2\n');
-        });
-    });
+        .then(result => {
+          expect(result).to.contain('interventionRef,totalinterested,totalapplied,totalinterview,' +
+            'totalfailure,totalsuccess,totalSaved\n');
+          expect(result).to.contain('\nFILTER,2,1,1,1,1,6\n');
+          expect(result).to.contain('\nRANDO,0,0,0,0,0,1\n');
+        })
+    );
   });
 });
